export functions in izh.c sended to writer.c
