+++ include/izh.c
@@ -859,77 +859,3 @@ int *gen_types(int num, double *ratio)
     return types;
 }
 
-
-void export_env(char tag[], network_info_t *info, int num_add, ...){
-    char fname[200];
-    sprintf(fname, "%s_env.json", tag);
-
-    JSON_Value *root_value;
-    JSON_Object *root_obj;
-
-    root_value = json_value_init_object();
-    root_obj = json_value_get_object(root_value);
-
-    json_object_set_number(root_obj, "num_cells", info->num_cells);
-    write_array_d(root_obj, "cell_type_ratio", info->cell_type_ratio, 2);
-    write_array_i(root_obj, "mean_degs, e->", info->mean_degs[0], 2);
-    write_array_i(root_obj, "mean_degs, i->", info->mean_degs[1], 2);
-    write_array_d(root_obj, "g_syn, e->", info->gsyns[0], 2);
-    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
-
-    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
-    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
-    
-    json_object_set_number(root_obj, "num_bck", info->num_bck);
-    json_object_set_number(root_obj, "fr_bck", info->frbck[0]);
-    write_array_d(root_obj, "p_bck", info->pbck[0], 2);
-    write_array_d(root_obj, "g_bck", info->gbck[0], 2);
-    json_object_set_number(root_obj, "dt", _dt);
-
-    json_object_set_number(root_obj, "t_delay_m", info->t_delay_m);
-    json_object_set_number(root_obj, "t_delay_s", info->t_delay_std);
-
-    // read additional parameters
-    va_list ap;
-    va_start(ap, num_add);
-    for (int n=num_add; n>0; n--)
-    {
-        char *var_name = va_arg(ap, char*);
-        double var = va_arg(ap, double);
-        json_object_set_number(root_obj, var_name, var);
-    }
-    va_end(ap);
-
-    json_serialize_to_file_pretty(root_value, fname);
-    json_value_free(root_value);
-}
-
-
-void export_ntk(syn_t *syns, char fname[])
-{
-    FILE *fp = fopen(fname, "w");
-    for (int n=0; n<syns->num_syns; n++){
-        fprintf(fp, "%d,%d,%d,%f,%f\n", n, syns->id_pre_neuron[n], syns->id_post_neuron[n], syns->weight[n], syns->veq[n]);
-    }
-    fclose(fp);
-}
-
-
-void write_array_d(JSON_Object *root_obj, char arr_name[], double arr1d[], int narr)
-{
-    json_object_set_value(root_obj, arr_name, json_value_init_array());
-    JSON_Array *arr = json_object_get_array(root_obj, arr_name);
-    for (int n=0; n<narr; n++){
-        json_array_append_number(arr, arr1d[n]);
-    }
-}
-
-
-void write_array_i(JSON_Object *root_obj, char arr_name[], int arr1d[], int narr)
-{
-    json_object_set_value(root_obj, arr_name, json_value_init_array());
-    JSON_Array *arr = json_object_get_array(root_obj, arr_name);
-    for (int n=0; n<narr; n++){
-        json_array_append_number(arr, arr1d[n]);
-    }
-}
