+++ include/writer.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdarg.h>
 #include "writer.h"
 
 
@@ -45,29 +46,6 @@ void write_result(writer_t *fid_obj, int nstep, neuron_t *cells)
     if (fid_obj->mod & 8){
         write_spike(fid_obj->ft_spk, nstep, cells->id_fired);
     }
-}
-
-
-void write_env(writer_t *fid_obj, int num_cells, int num_bck, double tmax, double dt, int *cell_types)
-{
-    char fname[200];
-    sprintf(fname, "%s_env.txt", fid_obj->tag);
-    FILE *fp = fopen(fname, "w");
-
-    fprintf(fp, "num_cells=%d\n", num_cells);
-    fprintf(fp, "num_bck=%d\n", num_bck);
-    fprintf(fp, "tmax=%f\n", tmax);
-    fprintf(fp, "dt=%f\n", dt);
-    for (int n=0; n<num_cells; n++){
-        fprintf(fp, "%d,", cell_types[n]);
-    }
-    // file list
-    fprintf(fp, "\n");
-    for (int i=1; i<16; i*=2){
-        fprintf(fp, "%d", is_opened(fid_obj, i));
-    }
-
-    fclose(fp);
 }
 
 
@@ -150,6 +128,105 @@ void write_spike_dat(writer_t *fid_obj, neuron_t *cells)
     fclose(fp);
 
     free(t_spk_flat);
+}
+
+
+void write_env(writer_t *fp_obj, network_info_t *info, int num_add, ...){
+    char fname[200];
+    sprintf(fname, "%s_env.json", fp_obj->tag);
+
+    JSON_Value *root_value;
+    JSON_Object *root_obj;
+
+    root_value = json_value_init_object();
+    root_obj = json_value_get_object(root_value);
+
+    json_object_set_number(root_obj, "num_cells", info->num_cells);
+    write_array_d(root_obj, "cell_type_ratio", info->cell_type_ratio, 2);
+    write_array_i(root_obj, "mean_degs, e->", info->mean_degs[0], 2);
+    write_array_i(root_obj, "mean_degs, i->", info->mean_degs[1], 2);
+    write_array_d(root_obj, "g_syn, e->", info->gsyns[0], 2);
+    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
+
+    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
+    write_array_d(root_obj, "g_syn, i->", info->gsyns[1], 2);
+    
+    json_object_set_number(root_obj, "num_bck", info->num_bck);
+    json_object_set_number(root_obj, "fr_bck", info->frbck[0]);
+    write_array_d(root_obj, "p_bck", info->pbck[0], 2);
+    write_array_d(root_obj, "g_bck", info->gbck[0], 2);
+    json_object_set_number(root_obj, "dt", _dt);
+
+    json_object_set_number(root_obj, "t_delay_m", info->t_delay_m);
+    json_object_set_number(root_obj, "t_delay_s", info->t_delay_std);
+
+    // read additional parameters
+    va_list ap;
+    va_start(ap, num_add);
+    for (int n=num_add; n>0; n--)
+    {
+        char *var_name = va_arg(ap, char*);
+        double var = va_arg(ap, double);
+        json_object_set_number(root_obj, var_name, var);
+    }
+    va_end(ap);
+
+    json_serialize_to_file_pretty(root_value, fname);
+    json_value_free(root_value);
+}
+
+/***
+void write_env(writer_t *fid_obj, int num_cells, int num_bck, double tmax, double dt, int *cell_types)
+{
+    char fname[200];
+    sprintf(fname, "%s_env.txt", fid_obj->tag);
+    FILE *fp = fopen(fname, "w");
+
+    fprintf(fp, "num_cells=%d\n", num_cells);
+    fprintf(fp, "num_bck=%d\n", num_bck);
+    fprintf(fp, "tmax=%f\n", tmax);
+    fprintf(fp, "dt=%f\n", dt);
+    for (int n=0; n<num_cells; n++){
+        fprintf(fp, "%d,", cell_types[n]);
+    }
+    // file list
+    fprintf(fp, "\n");
+    for (int i=1; i<16; i*=2){
+        fprintf(fp, "%d", is_opened(fid_obj, i));
+    }
+
+    fclose(fp);
+}
+***/
+
+
+void write_ntk(char fname[], syn_t *syns)
+{
+    FILE *fp = fopen(fname, "w");
+    for (int n=0; n<syns->num_syns; n++){
+        fprintf(fp, "%d,%d,%d,%f,%f\n", n, syns->id_pre_neuron[n], syns->id_post_neuron[n], syns->weight[n], syns->veq[n]);
+    }
+    fclose(fp);
+}
+
+
+void write_array_d(JSON_Object *root_obj, char arr_name[], double arr1d[], int narr)
+{
+    json_object_set_value(root_obj, arr_name, json_value_init_array());
+    JSON_Array *arr = json_object_get_array(root_obj, arr_name);
+    for (int n=0; n<narr; n++){
+        json_array_append_number(arr, arr1d[n]);
+    }
+}
+
+
+void write_array_i(JSON_Object *root_obj, char arr_name[], int arr1d[], int narr)
+{
+    json_object_set_value(root_obj, arr_name, json_value_init_array());
+    JSON_Array *arr = json_object_get_array(root_obj, arr_name);
+    for (int n=0; n<narr; n++){
+        json_array_append_number(arr, arr1d[n]);
+    }
 }
