+++ include/writer.h
@@ -2,6 +2,7 @@
 #define WRITER
 
 #include "izh.h"
+#include "parson.h"
 
 typedef enum _WRITER_VAR
 {
@@ -10,6 +11,7 @@ typedef enum _WRITER_VAR
     I_ONLY = 4,
     SPK_ONLY = 8,
     SPK_DAT = 16
+    
 } WRITER_VAR;
 
 
@@ -25,9 +27,9 @@ typedef struct _writer_t
 } writer_t;
 
 void init_writer(writer_t *fid_obj, char tag[], WRITER_VAR mod);
-void write_env(writer_t *fid_obj, int  num_cells, int num_bck, double tmax, double dt, int *cell_types);
+void write_env(writer_t *fp_obj, network_info_t *info, int num_add, ...);
+// void write_env(writer_t *fid_obj, int  num_cells, int num_bck, double tmax, double dt, int *cell_types);
 void write_result(writer_t *fid_obj, int nstep, neuron_t *cells);
-// void write(int nstep);
 void write_data(FILE *fp, int num_x, double *x);
 void write_spike(FILE *fp, int nstep, int *t_spk);
 void write_spike_dat(writer_t *fid_obj, neuron_t *cells);
@@ -35,5 +37,8 @@ void end_writer(writer_t *fid_obj);
 FILE *open_file(char fname[], char *type);
 void close_file(FILE *fid);
 int is_opened(writer_t *fid_obj, int mod);
+void write_array_d(JSON_Object *root_obj, char arr_name[], double arr1d[], int narr);
+void write_array_i(JSON_Object *root_obj, char arr_name[], int arr1d[], int narr);
+void write_ntk(char fname[], syn_t *syns);
 
 #endif
